Add files via upload
Draft1 of Texture Synthesis. This file was not meant for amazing results, rather this file was meant for me to understand how to go about training this with Tensorflow. I am going to be running more trials with different pictures and different tricks added on top of what is here.