Add files via upload
Used Texture Synthesis Original paper to do this work. This was on my own and turned out to give some pretty cool results. The main focus of the paper is the fact that we can use a pretrained object detection model to extract the feature statistics of the intermediate convolution layers. Here I skipped some of the convolution layers but I am sure incorporating these in as well would actually drastically improve the results. The statistics are captured via a gram matrix of the convolution layers where we vectorize the height and width and each channel has 1 row corresponding to it, leading to an NxC Matrix where N = Height*Width and C = Channels, we then flip it so its CxN then get the gram matrix which is just the matrix multiplied by its own transpose. This captures feature statistics but also conveniently makes it so the output size is only dependant on the channel amount so no matter the size of the height and width we can capture the euclidean distance between both the noise and target texture. Then after this I simply normalize the euclidean distance. Then run a loop and update the picture using gradient descent. Note that in order for GradientTape() to work, the noise variable (image we are updating) has to be a direct input into the function (under where GradientTape() is deployed) (For example, gradient(loss(x), z) where x = f(z) will not work, but gradient(loss(z), z) does work! When we are trying to update z in tensorflow) otherwise the gradient feature won't work. I learned this the hard way and sat here for about 3 hours figuring that simple fact out which is not really very clear in documentation or error messages. But overall I am glad I was able to complete this on my own. Fun project!